Add method to get event source ID to Event
Now, user must check the event is whether `user` or `group` or `event`.  It is redundant for me to write such code everywhere, so created intelligent getter method to do that.
`#getEventSourceId()` returns the identifier of the event source that associated with event source type (i.e. userId, groupId or roomId).