Support connection checking from dashboard
Connection checker on LINE Developers dashboard will fail after copy and paste this code.
Because the checker sends to the server as a "sticker" type message, but now only "text" type messages are supported.

I don't know why LINE send as a "sticker" type and how we detect well that a message is for doing connection check. But it works. :P

[![Image from Gyazo](https://i.gyazo.com/a65f074868a282712b83aa632dc07e9b.png)](https://gyazo.com/a65f074868a282712b83aa632dc07e9b)