Bump up openapi generator version to 7.11.0 (#667)
Resolve https://github.com/line/line-bot-sdk-php/issues/636
Resolve https://github.com/line/line-bot-sdk-php/issues/625

The version of the OpenAPI generator used internally has been updated
from 7.10.0 to 7.11.0.

I've changed the version of the OpenAPI generator in
https://github.com/line/line-bot-sdk-php/commit/863e54667467e40a48465c4f105b75dba845fbbb
and generated code and documentation in
https://github.com/line/line-bot-sdk-php/commit/44b520b9ca96851eff19f329dd1f4380f2e1361f
and
https://github.com/line/line-bot-sdk-php/commit/7955fdf10e27e9ee033b62985b98e1612848a11.

The main change is that it now explicitly describes nullable types like
this.

```diff
- function foo (array $data = null)
+ function foo (array|null $data = null)
```

This should allow support for PHP 8.4.